+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/TabPane/CustomDisplayTabPane.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames/bind';
 import IconSearch from 'terra-icon/lib/icon/IconSearch';
-import TabPane from '../../../../TabPane';
+import TabPane from 'terra-tabs';
 import styles from './common/TabPane.test.module.scss';
 
 const cx = classNames.bind(styles);